% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-getData.R
\name{getData}
\alias{getData}
\title{API wrapper that fetches data from IsoMemo Applications}
\usage{
getData(
  db = getDatabaseList(mappingId = "IsoMemo"),
  category = NULL,
  field = NULL,
  mappingId = "IsoMemo"
)
}
\arguments{
\item{db}{database options:  "14CSea"   "CIMA"     "IntChron" "LiVES"}

\item{category}{domain specific categories of fields to retrieve: "Dating info","Isotopic proxies." If set to NULL (default) all categories are returned}

\item{field}{fields to return. If set to NULL (default) all fields will be returned}

\item{mappingId}{(character) Optionally, provide a specific mappingId in order to obtain a list
of databases only for that mapping. Check available mapping ids with getMappingIds().}
}
\value{
A data frame containing the requested databases, category domains, and variables of interest from the user
}
\description{
Data retrieval for databases within the IsoMemo initiative (https://isomemo.com/). Given the parameters Database
name, category names, and field names.
}
\examples{
getData(db = "IntChron")
getData(db = "IntChron", category = "Location")
getData(db = "IntChron", category = "Location", field = "latitude")
getData(db = "IntChron", category = "Location", field = "latitude", mappingId = "IsoMemo")

}
