% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-getData.R
\name{getFields}
\alias{getFields}
\title{Get field mapping table}
\usage{
getFields()
}
\value{
A dataframe that contains the requested field, type, and categories of data specs
}
\description{
Get field mapping table
}
